refactor: move select items to ui subpackage
