+++ disnake/ui/select/__init__.py
@@ -1,0 +1,13 @@
+# SPDX-License-Identifier: MIT
+
+"""
+disnake.ui.select
+~~~~~~~~~~~~~~~~~~
+
+Select Menu UI Kit Types
+
+:copyright: (c) 2021-present Disnake Development
+:license: MIT, see LICENSE for more details.
+"""
+from .base import *
+from .string import *
